feat : boardService TestCode

+++ src/main/java/com/example/basicboard/domain/board/service/BoardService.java
@@ -3,7 +3,9 @@ package com.example.basicboard.domain.board.service;
 
 import com.example.basicboard.domain.board.Entity.Board;
 import com.example.basicboard.domain.board.dto.BoardRequestDTO;
+import com.example.basicboard.domain.board.dto.BoardResponseDTO;
 import com.example.basicboard.domain.board.repository.BoardRepository;
+import com.example.basicboard.domain.member.entity.Member;
 import com.example.basicboard.global.message.Message;
 import com.example.basicboard.global.message.StatusEnum;
 import jakarta.persistence.EntityNotFoundException;
@@ -14,6 +16,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.ArrayList;
 import java.util.List;
 
 @Service
@@ -25,57 +28,71 @@ public class BoardService {
 
 
     //게시글 전체조회
+
     @Transactional(readOnly = true)
-    public ResponseEntity<Message> readAll() {
-
+    public ResponseEntity<Message<List<BoardResponseDTO>>> readAll() {
         List<Board> boards = boardRepository.findAll();
-
-        Message message = Message.setSuccess(StatusEnum.OK, "게시글 전체조회", boards);
-
+        List<BoardResponseDTO> boardResponseDTO = new ArrayList<>();
+        for (Board board : boards) {
+            boardResponseDTO.add(new BoardResponseDTO(board));
+        }
+        Message<List<BoardResponseDTO>> message = Message.setSuccess(StatusEnum.OK, "게시글 전체조회", boardResponseDTO);
         return new ResponseEntity<>(message, HttpStatus.OK);
     }
 
     //게시글작성
     @Transactional
-    public ResponseEntity<Message> boardCreate(BoardRequestDTO requestDto) {
+    public ResponseEntity<Message<BoardResponseDTO>> boardCreate(BoardRequestDTO requestDto, Member member) {
 
         Board board = Board.builder()
                 .title(requestDto.getTitle())
                 .content(requestDto.getContent())
                 .date(requestDto.getDate())
+                .member(member)
                 .build();
         boardRepository.save(board);
+        BoardResponseDTO boardResponseDTO = new BoardResponseDTO(board);
 
-        Message message = Message.setSuccess(StatusEnum.OK, "게시글 작성성공", board);
+        Message<BoardResponseDTO> message = Message.setSuccess(StatusEnum.OK, "게시글 작성성공", boardResponseDTO);
         return new ResponseEntity<>(message, HttpStatus.OK);
 
     }
 
     //게시글 수정
     @Transactional
-    public ResponseEntity<Message> boardUpdate(Long id, BoardRequestDTO requestDto) {
-
+    public ResponseEntity<Message<BoardResponseDTO>> boardUpdate(Long id, BoardRequestDTO requestDto, Member member) {
         validateBoard(requestDto);
 
         Board board = findBoardById(id);
 
+
+        if (!member.getId().equals(board.getMember().getId())) {
+            System.out.println("같지않음");
+            return null;
+        }
+
         updateBoardFields(board, requestDto);
 
+        BoardResponseDTO boardResponseDTO = new BoardResponseDTO(board);
 
         boardRepository.save(board);
-
-        Message message = Message.setSuccess(StatusEnum.OK, "게시글 수정성공", board);
+        Message<BoardResponseDTO> message = Message.setSuccess(StatusEnum.OK, "게시글 수정성공", boardResponseDTO);
         return new ResponseEntity<>(message, HttpStatus.OK);
 
     }
 
     @Transactional
-    public ResponseEntity<Message> boardDelete(Long id) {
+    public ResponseEntity<Message<String>> boardDelete(Long boardId, Member member) {
+        Board board = findBoardById(boardId);
 
-        Board board = findBoardById(id);
+
+        if (!member.getId().equals(board.getMember().getId())) {
+            System.out.println("같지않음");
+            return null;
+        }
         boardRepository.delete(board);
 
-        Message message = Message.setSuccess(StatusEnum.OK, "게시글 삭제성공");
+        Message<String> message = Message.setSuccess(StatusEnum.OK, "게시글 삭제성공", "게시글 삭제성공");
         return new ResponseEntity<>(message, HttpStatus.OK);
 
     }
@@ -100,6 +117,12 @@ public class BoardService {
         }
         if (requestDto.getDate() != null) {
             board.setDate(requestDto.getDate());
+        }
+        if (requestDto.getAuthor() != null) {
+            board.setAuthor(requestDto.getAuthor());
+        }
+        if (requestDto.getHits() != null) {
+            board.setHits(requestDto.getHits());
         }
     }
